ManagedBuffer incr/decrRef helpers
