+++ include/Pothos/Framework/ManagedBuffer.hpp
@@ -115,6 +115,8 @@ public:
 
 private:
     friend BufferChunk;
+    void _incrRef(void);
+    void _decrRef(void);
     struct Impl; Impl *_impl;
     ManagedBuffer(Impl *impl);
     POTHOS_API friend bool operator==(const ManagedBuffer &lhs, const ManagedBuffer &rhs);
