+++ lib/Framework/ManagedBuffer.cpp
@@ -1,10 +1,26 @@
 // Copyright (c) 2013-2016 Josh Blum
 // SPDX-License-Identifier: BSL-1.0
 
-#include "ManagedBufferImpl.hpp"
 #include <Pothos/Framework/ManagedBuffer.hpp>
 #include <Pothos/Framework/BufferManager.hpp>
+#include <atomic>
 #include <cassert>
+
+struct Pothos::ManagedBuffer::Impl
+{
+    Impl(void):
+        counter(1),
+        slabIndex(0),
+        nextBuffer(nullptr)
+    {
+        return;
+    }
+    std::atomic<int> counter;
+    std::weak_ptr<BufferManager> weakManager;
+    SharedBuffer buffer;
+    size_t slabIndex;
+    Pothos::ManagedBuffer::Impl *nextBuffer;
+};
 
 Pothos::ManagedBuffer::ManagedBuffer(void):
     _impl(nullptr)
@@ -15,21 +31,12 @@ Pothos::ManagedBuffer::ManagedBuffer(void):
 Pothos::ManagedBuffer::ManagedBuffer(Impl *impl):
     _impl(impl)
 {
-    if (_impl != nullptr) _impl->counter++;
+    _incrRef();
 }
 
 void Pothos::ManagedBuffer::reset(void)
 {
-    //decrement the counter, and handle the last ref case
-    if (_impl != nullptr and _impl->counter.fetch_sub(1) == 1)
-    {
-        //there is a manager to push to, otherwise delete
-        std::shared_ptr<BufferManager> manager = _impl->weakManager.lock();
-        if (manager) manager->pushExternal(*this);
-        else delete _impl;
-    }
-
-    //zero out the class members
+    _decrRef();
     _impl = nullptr;
 }
 
@@ -43,13 +50,13 @@ void Pothos::ManagedBuffer::reset(BufferManager::Sptr manager, const SharedBuffe
 
 Pothos::ManagedBuffer::~ManagedBuffer(void)
 {
-    this->reset();
+    _decrRef();
 }
 
 Pothos::ManagedBuffer::ManagedBuffer(const ManagedBuffer &obj):
     _impl(obj._impl)
 {
-    if (_impl != nullptr) _impl->counter++;
+    _incrRef();
 }
 
 Pothos::ManagedBuffer::ManagedBuffer(ManagedBuffer &&obj):
@@ -60,15 +67,15 @@ Pothos::ManagedBuffer::ManagedBuffer(ManagedBuffer &&obj):
 
 Pothos::ManagedBuffer &Pothos::ManagedBuffer::operator=(const ManagedBuffer &obj)
 {
-    this->reset();
+    _decrRef();
     this->_impl = obj._impl;
-    if (_impl != nullptr) _impl->counter++;
+    _incrRef();
     return *this;
 }
 
 Pothos::ManagedBuffer &Pothos::ManagedBuffer::operator=(ManagedBuffer &&obj)
 {
-    this->reset();
+    _decrRef();
     this->_impl = obj._impl;
     obj._impl = nullptr;
     return *this;
@@ -108,4 +115,21 @@ Pothos::ManagedBuffer Pothos::ManagedBuffer::getNextBuffer(void) const
 {
     assert(*this);
     return Pothos::ManagedBuffer(_impl->nextBuffer);
+}
+
+void Pothos::ManagedBuffer::_incrRef(void)
+{
+    if (_impl != nullptr) _impl->counter++;
+}
+
+void Pothos::ManagedBuffer::_decrRef(void)
+{
+    //decrement the counter, and handle the last ref case
+    if (_impl != nullptr and _impl->counter.fetch_sub(1) == 1)
+    {
+        //there is a manager to push to, otherwise delete
+        std::shared_ptr<BufferManager> manager = _impl->weakManager.lock();
+        if (manager) manager->pushExternal(*this);
+        else delete _impl;
+    }
 }
